user entity, user repository, gitignore env 추가
