--- src/main/java/com/goormitrip/goormitrip/MainController.java
@@ -1,19 +1,0 @@
-package com.goormitrip.goormitrip;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class MainController {
-
-    @GetMapping("/")
-    public String home(Model model) {
-        return "main";
-    }
-
-    @GetMapping("/signup")
-    public String signupForm() {
-        return "signup";
-    }
-}
