+++ src/main/java/com/goormitrip/goormitrip/controller/MainController.java
@@ -1,0 +1,33 @@
+package com.goormitrip.goormitrip.controller;
+
+import com.goormitrip.goormitrip.domain.User;
+import com.goormitrip.goormitrip.repository.UserRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+
+@Controller
+public class MainController {
+
+    @Autowired
+    private UserRepository userRepository;
+
+    @GetMapping("/")
+    public String home(Model model) {
+        return "main";
+    }
+
+    @GetMapping("/signup")
+    public String signupForm() {
+        return "signup";
+    }
+
+    @PostMapping("/signup")
+    public String processSignup(User user) {
+        userRepository.save(user);
+        return "redirect:/login";
+    }
+
+}
