+++ src/main/java/com/goormitrip/goormitrip/domain/User.java
@@ -1,0 +1,26 @@
+package com.goormitrip.goormitrip.domain;
+
+import jakarta.persistence.*;
+import lombok.Data;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDateTime;
+
+@Data
+@Entity
+public class User {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(unique = true)
+    private String email;
+    private String password;
+    private String phone;
+
+    @CreationTimestamp
+    private LocalDateTime createAt;
+}
+
+
