+++ src/main/resources/application.properties
@@ -1,8 +1,10 @@
 spring.application.name=goormitrip
 
+spring.config.import=optional:file:.env[.properties]
+
 spring.datasource.url=jdbc:mysql://localhost:3306/goormitrip
-spring.datasource.username=root
-spring.datasource.password=your_password
+spring.datasource.username=${DB_USERNAME}
+spring.datasource.password=${DB_PASSWORD}
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 
 spring.jpa.hibernate.ddl-auto=update
